requirements.txt
altair
asttokens
attrs
backcall
backports.zoneinfo
blinker
Brotli
buil==1.0.3
cachetools
certifi
charset-normalizer
click
colorama
comm
contourpy==1.1.1
cycler
debugpy
decorator
docutils==0.20.1
executing
filelock
fonttools==4.45.1
fsspec==2023.10.0
gitdb
GitPython
idna
importlib-metadata
importlib-resources
ipykernel
ipython
ipywidgets
jaraco.classes==3.3.0
jedi
Jinja2
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
jupyterlab-widgets
keyring==24.3.0
kiwisolver
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
more-itertools==10.1.0
mpmath
nest-asyncio
networkx
nh3==0.2.15
numpy
opencv-python==4.8.1.78
packaging
pandas
parso
pickleshare
Pillow
pkginfo==1.9.6
pkgutil_resolve_name
platformdirs
prompt-toolkit
protobuf
psutil
pure-eval
py-cpuinfo==9.0.0
pyarrow==11.0.0
pydeck
Pygments
pyparsing
pyproject_hooks==1.0.0
PySide6==6.6.1
PySide6-Addons==6.6.1
PySide6-Essentials==6.6.1
PySocks
python-dateutil
pytz
pywin32==306
pywin32-ctypes==0.2.2
PyYAML
pyzmq
readme-renderer==42.0
referencing
requests
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich
rpds-py
scipy==1.10.1
seaborn==0.13.0
shiboken6==6.6.1
six
smmap
stack-data
streamlit
supervision
sympy
tenacity
thop==0.1.1.post2209072238
toml
tomli==2.0.1
toolz
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tornado
tqdm==4.66.1
traitlets
twine==4.0.2
typing_extensions
tzdata
tzlocal
ultralytics==8.0.227
ultralytics-yolo==0.0.1
urllib3
validators
watchdog
wcwidth
widgetsnbextension
win-inet-pton
zipp
